clauditor-24h.8: US-008 — create cli/ package + extract small commands
Refactor src/clauditor/cli.py (~2716 lines) into a cli/ subpackage with
small commands moved into their own modules. Grade, compare, audit,
extract, triggers, and suggest stay in cli/__init__.py for US-009/US-010.

Layout:
- git mv cli.py -> cli/__init__.py preserves blame for remaining commands
- cli/validate.py: cmd_validate + add_parser (lazy-imports shared helpers)
- cli/run.py:      cmd_run + add_parser
- cli/init.py:     cmd_init + add_parser
- cli/doctor.py:   cmd_doctor + add_parser
- cli/capture.py:  cmd_capture + add_parser
- cli/trend.py:    cmd_trend + add_parser (carries _positive_int copy)

Each module exposes add_parser(subparsers) -> None and
cmd_<name>(args) -> int. cli/__init__.py imports each module AFTER all
shared helpers are defined so lazy imports inside cmd_validate resolve
cleanly. Handler symbols are re-exported at the package level so
existing test imports (from clauditor.cli import cmd_run) keep working.

Shared helpers (_load_spec_or_report, _write_run_dir,
_print_failing_transcript_slice, _append_validate_history,
_relative_to_repo, _unit_float, _positive_int) stay in cli/__init__.py;
cmd_validate imports them lazily inside the function body to avoid a
circular import.

Test adjustments (patch-path updates only, no behavior changes):
- clauditor.cli.SkillRunner -> clauditor.cli.run.SkillRunner (TestCmdRun)
- clauditor.cli.SkillRunner -> clauditor.cli.capture.SkillRunner (TestCmdCapture)
- clauditor.cli.allocate_iteration -> clauditor.cli.validate.allocate_iteration
- clauditor.cli.run_assertions -> clauditor.cli.validate.run_assertions
  (only in tests that verify cmd_validate internals)

Help-output diff: all 12 subcommands + top-level --help produce
byte-identical output to pre-refactor baselines.

cli.py line delta:
- Before:  cli.py = 2716 lines
- After:   cli/__init__.py = 2173 lines (-543)
- Extract: 674 lines across 6 new modules (run/init/doctor/capture/trend/validate)

Quality gates:
- ruff: clean
- pytest: 1171 passed, 0 failed
- coverage: 96.18% (gate: 80%)

Entry point unchanged: clauditor = "clauditor.cli:main" resolves via
the new cli/ package's __init__.py.
