+++ src/clauditor/cli/doctor.py
@@ -1,0 +1,119 @@
+"""``clauditor doctor`` — environment diagnostics."""
+
+from __future__ import annotations
+
+import argparse
+import sys
+from pathlib import Path
+
+
+def add_parser(subparsers: argparse._SubParsersAction) -> None:
+    """Register the ``doctor`` subparser."""
+    subparsers.add_parser(
+        "doctor",
+        help=(
+            "Report environment diagnostics "
+            "(Python, SDK, claude CLI, plugin, install)"
+        ),
+    )
+
+
+def cmd_doctor(args: argparse.Namespace) -> int:
+    """Read-only environment diagnostics (DEC-005/008/014).
+
+    Always exits 0 — this is a reporting tool, not a CI gate.
+    """
+    import importlib.metadata
+    import importlib.util
+    import shutil
+
+    checks: list[tuple[str, str, str]] = []
+
+    py_version = sys.version_info
+    if py_version >= (3, 11):
+        checks.append(
+            (
+                "python",
+                "ok",
+                f"Python {py_version.major}.{py_version.minor}.{py_version.micro}",
+            )
+        )
+    else:
+        checks.append(
+            (
+                "python",
+                "fail",
+                f"Python {py_version.major}.{py_version.minor} < 3.11 (required)",
+            )
+        )
+
+    if importlib.util.find_spec("anthropic") is not None:
+        checks.append(("anthropic", "ok", "SDK importable"))
+    else:
+        checks.append(
+            (
+                "anthropic",
+                "warn",
+                "SDK not installed (required only for Layer 2/3 grading)",
+            )
+        )
+
+    claude_path = shutil.which("claude")
+    if claude_path:
+        checks.append(("claude-cli", "ok", claude_path))
+    else:
+        checks.append(
+            ("claude-cli", "fail", "`claude` not found on PATH")
+        )
+
+    try:
+        # Version-agnostic lookup: `entry_points(group=...)` is 3.10+, but
+        # `doctor` must keep working even when the Python-version check
+        # itself is about to fail, so fall back to filtering manually.
+        eps = importlib.metadata.entry_points()
+        if hasattr(eps, "select"):
+            eps = eps.select(group="pytest11")
+        else:
+            eps = [
+                ep for ep in eps
+                if getattr(ep, "group", None) == "pytest11"
+            ]
+        names = [ep.name for ep in eps]
+        if "clauditor" in names:
+            checks.append(
+                ("pytest-plugin", "ok", "clauditor registered under pytest11")
+            )
+        else:
+            checks.append(
+                (
+                    "pytest-plugin",
+                    "fail",
+                    f"clauditor not registered (found: {names})",
+                )
+            )
+    except Exception as e:  # pragma: no cover - defensive
+        checks.append(("pytest-plugin", "fail", f"entry_points lookup failed: {e}"))
+
+    spec = importlib.util.find_spec("clauditor")
+    if spec is not None and spec.origin is not None:
+        origin = Path(spec.origin).resolve()
+        if "site-packages" in origin.parts and not origin.is_symlink():
+            checks.append(
+                (
+                    "editable-install",
+                    "warn",
+                    f"clauditor installed non-editable at {origin.parent} "
+                    f"— source edits will not propagate",
+                )
+            )
+        else:
+            checks.append(("editable-install", "ok", str(origin.parent)))
+    else:
+        checks.append(("editable-install", "fail", "clauditor package not importable"))
+
+    width = max(len(name) for name, _, _ in checks)
+    for name, status, detail in checks:
+        tag = f"[{status}]"
+        print(f"{tag:<7} {name:<{width}}  {detail}")
+
+    return 0
