+++ src/clauditor/cli/run.py
@@ -1,0 +1,35 @@
+"""``clauditor run`` — run a skill and print its output."""
+
+from __future__ import annotations
+
+import argparse
+import sys
+from pathlib import Path
+
+from clauditor.runner import SkillRunner
+
+
+def add_parser(subparsers: argparse._SubParsersAction) -> None:
+    """Register the ``run`` subparser."""
+    p_run = subparsers.add_parser("run", help="Run a skill and print output")
+    p_run.add_argument("skill", help="Skill name (e.g., find-kid-activities)")
+    p_run.add_argument("--args", help="Arguments to pass to the skill")
+    p_run.add_argument("--project-dir", help="Project directory (default: cwd)")
+    p_run.add_argument("--timeout", type=int, default=180, help="Timeout in seconds")
+
+
+def cmd_run(args: argparse.Namespace) -> int:
+    """Run a skill and print its output."""
+    runner = SkillRunner(
+        project_dir=args.project_dir or Path.cwd(),
+        timeout=args.timeout,
+    )
+    result = runner.run(args.skill, args.args or "")
+
+    if result.error:
+        print(f"ERROR: {result.error}", file=sys.stderr)
+
+    if result.output:
+        print(result.output)
+
+    return result.exit_code
